+++ packages/isomorphic/src/ripemd160/index.ts
@@ -1,0 +1,4 @@
+import { createHash } from 'crypto'
+import wrapCryptoCreateHash from '../internal/wrapCryptoCreateHash'
+
+export const ripemd160 = wrapCryptoCreateHash('ripemd160', createHash)