+++ packages/isomorphic/src/sha256/browser.ts
@@ -1,0 +1,5 @@
+import { sha256 as nobleImpl } from '@noble/hashes/sha256'
+
+import wrapNoble from '../internal/wrapNoble'
+
+export const sha256 = wrapNoble(nobleImpl)