+++ packages/isomorphic/src/sha256/index.ts
@@ -1,0 +1,4 @@
+import { createHash } from 'crypto'
+import wrapCryptoCreateHash from '../internal/wrapCryptoCreateHash'
+
+export const sha256 = wrapCryptoCreateHash('sha256', createHash)