+++ packages/isomorphic/src/sha512/index.ts
@@ -1,0 +1,4 @@
+import { createHash } from 'crypto'
+import wrapCryptoCreateHash from '../internal/wrapCryptoCreateHash'
+
+export const sha512 = wrapCryptoCreateHash('sha512', createHash)