+++ packages/ripple-binary-codec/src/hashes.ts
@@ -2,7 +2,7 @@ import { HashPrefix } from './hash-prefixes'
 import { Hash256 } from './types'
 import { BytesList } from './serdes/binary-serializer'
 import { Buffer } from 'buffer/'
-import { sha512 } from '@xrpl/crypto/sha512'
+import { sha512 } from '@xrplf/isomorphic/sha512'
 
 /**
  * Class for hashing with SHA512