+++ packages/xrpl-crypto/jest.config.js
@@ -1,8 +1,0 @@
-// Jest configuration for api
-const base = require('../../jest.config.base.js')
-
-module.exports = {
-  ...base,
-  roots: [...base.roots, '<rootDir>/test'],
-  displayName: '@xrpl/crypto',
-}