+++ packages/xrpl-crypto/src/sha256/browser.ts
@@ -1,6 +1,0 @@
-import { sha256 as nobleImpl } from '@noble/hashes/sha256'
-
-import wrapNoble from '../internal/wrapNoble'
-
-// eslint-disable-next-line import/prefer-default-export
-export const sha256 = wrapNoble(nobleImpl)