+++ packages/xrpl-crypto/src/sha512/browser.ts
@@ -1,6 +1,0 @@
-import { sha512 as nobleImpl } from '@noble/hashes/sha512'
-
-import wrapNoble from '../internal/wrapNoble'
-
-// eslint-disable-next-line import/prefer-default-export
-export const sha512 = wrapNoble(nobleImpl)