+++ packages/xrpl/src/utils/hashes/sha512Half.ts
@@ -1,5 +1,5 @@
-import { sha512 } from '@xrpl/crypto/sha512'
-import { bytesToHex, hexToBytes } from '@xrpl/crypto/utils'
+import { sha512 } from '@xrplf/isomorphic/sha512'
+import { bytesToHex, hexToBytes } from '@xrplf/isomorphic/utils'
 
 const HASH_BYTES = 32